Problem testing, redux.
This is my (hopefully last ;-)) attempt to make problem testing
reasonable. It adds a `Problems` type which is just `[]Problem`,
and gives it an `Equal` method so that it can be passed to
`cmp.Diff`.

There is also a `SetDescriptor` helper used for tests.

As before, I have migrated only a couple of tests to get feedback
before migrating the rest.

And hopefully this is the last major interface change. :-)