[feat] Add helper method for proto3 templates.
This adds a `testutils.ParseProto3String` and
`testutils.ParseProto3Tmpl` method, to make it easier to
parse protobuf string templates in tests, and keep the focus
on what is being tested.

Each method adds `syntax = "proto3";` to the top (we basically
only work in proto3 so it is silly to boilerplate it over and over)
and the template method also parses out a template.