[feat] Linter documentation in GitHub Pages.
This (incredibly large) PR adds a `docs/` directory that we can
serve with GitHub Pages.

It provides documentation for three linter rules; the remainder
will come in follow-up PRs.

Note that as part of this, I intend to enforce (and soon lint for)
some consistency on linter rules, including:

- One rule per file.
- The rule's filename must be derived from the rule name
  (converting `-` to `_` but no other changes).

This is to make things like "View Implmenetation" easy.