Don't enforce AIP-0123 on Response messages
In Cloud KMS, we echo the name of the resource used for cryptographic
operations in response messages, so that callers can verify that the
correct resource was used.

(Corruption of a single bit in the `name` field of a request message
could end up with the wrong key being used. We'd like for a caller to be
able to detect when that's happened.)