feat: Enforce OUTPUT_ONLY behavior for State fields
google.aip.dev/216 calls for `State` fields to be
OUTPUT_ONLY.

Add a rule that requires OUTPUT_ONLY for fields which have
an enum type ending with `State`.

Note that the AIP does not mention rules for the *field name* for state
fields, we ignore the field name in the linter.