chore(deps): downgrade and freeze protoreflect (#1449)
As per conversation in #1393, we need to downgrade `jhump/protoreflect` to v1.16.0, because v1.17.0 has packages critical to api-linter marked as deprecated. We cannot move off of the packages deprecated in that release in a timely manner because it would require a major revision of the public api-linter surface, which requires much more planning. As such, we will freeze our minimum dep at v1.16.0, and dependents can upgrade to v1.17.0 locally if they'd like to.

Also deleting the staticcheck configuration that I added to disable the deprecated dependency warning locally as we again want its protection.