fix(AIP-203): field-behavior-required ignore imported request types (#1504)
fixes a bug where the method input type used as the starting point for the `field-behavior-required` recursion did not check if the input type itself was from the same package. It is not a super common practice in general, but appears a lot for APIs that mix in the google.iam.v1 IAM Policy service RPCs and reuse their request messages. Same thing could happen with `google.api.HttpBody` which is sometimes used as a RPC input type.

Updates #1503