fix(AIP-123): handle errant slash in pattern parsing (#1517)
When parsing the resource name `pattern`s to evaluate if a resource is a root/top-level resource, and pull out the parent resource if not, there was a bug where a leading slash on a root/top-level resource would make it think the pattern was multi-segmented (because it just checks for how many segments separators the `pattern` has). This would result in following through to logic that expected multiple items in a slice when there was only one, resulting in an out of bounds panic. We can handle this by trimming leading/trailing separators from the pattern before evaluating the pattern as a root/top-level resource 

Updates #1514