chore(ci): build release assets from trigger tag (#1528)
Now that we have a `v2` branch that we want to make a release candidate from, we need to ensure that if we tag a release from that branch, we build release assets from that tag specifically - not from `main`. This changes the `checkout` action to specifically checkout the release tag prior to building anything.

This is also a general improvement for the `main` branch as it ensures the assets we build/upload are from the exact same tag rather than whatever is in `main` when the release was created (which may be after the tag was created in non-Release-Please cases, which would be exceptional).